added ssl losses
