+++ pl_bolts/models/self_supervised/amdim/amdim_module.py
@@ -7,7 +7,7 @@ import torch.optim as optim
 from torch.optim.lr_scheduler import MultiStepLR
 from torch.utils.data import DataLoader
 
-from pl_bolts.metrics.self_supervised.losses import AMDIMLossNCE
+from pl_bolts.losses.self_supervised_learning import AMDIMLossNCE
 from pl_bolts.models.self_supervised.amdim.datasets import AMDIMPretraining
 from pl_bolts.models.self_supervised.amdim.networks import AMDIMEncoder
